linfa_kernel

Struct KernelBase

pub struct KernelBase<K1: Inner, K2: Inner>
where
    K1::Elem: Float,
    K2::Elem: Float,
{
    pub inner: KernelInner<K1, K2>,
    pub method: KernelMethod<K1::Elem>,
}

A generic kernel

Fields
inner: KernelInner<K1, K2>method: KernelMethod<K1::Elem>

The inner product that will be used by the kernel

Implementations

impl<F: Float, K1: Inner<Elem = F>, K2: Inner<Elem = F>> KernelBase<K1, K2>

pub fn is_linear(&self) -> bool

Whether the kernel is a linear kernel

Returns

true: if the kernel is linear

false: otherwise

pub fn params() -> KernelParams<F, CommonNearestNeighbour>

Generates the default set of parameters for building a kernel. Use this to initialize a set of parameters to be customized using KernelParams’s methods

pub fn params_with_nn<N: NearestNeighbour>(nn_algo: N) -> KernelParams<F, N>

Generate parameters with a specific nearest neighbour algorithm

pub fn dot(&self, rhs: &ArrayView2<'_, F>) -> Array2<F>

Performs the matrix product between the kernel matrix and the input

Parameters

rhs: The matrix on the right-hand side of the multiplication

Returns

A new matrix containing the matrix product between the kernel and rhs

Panics

If the shapes of kernel and rhs are not compatible for multiplication

pub fn sum(&self) -> Array1<F>

Sums all elements in the same row of the kernel matrix

Returns

A new array with the sum of all the elements in each row

pub fn size(&self) -> usize

Gives the size of the side of the square kernel matrix

pub fn column(&self, i: usize) -> Vec<F>

Getter for a column of the kernel matrix

Params

i: the index of the column

Returns

The i-th column of the kernel matrix, stored as a Vec

Panics

If i is out of bounds

pub fn to_upper_triangle(&self) -> Vec<F>

Getter for the data in the upper triangle of the kernel matrix

Returns

A copy of all elements in the upper triangle of the kernel matrix, stored in a Vec

pub fn diagonal(&self) -> Array1<F>

Getter for the elements in the diagonal of the kernel matrix

Returns

A new array containing the copy of all elements in the diagonal fo the kernel matrix

impl<'a, F: Float> KernelBase<ArrayBase<OwnedRepr<F>, Dim<[usize; 2]>>, CsMatBase<F, usize, Vec<usize>, Vec<usize>, Vec<F>>>

pub fn new<N: NearestNeighbour>( dataset: ArrayView2<'a, F>, params: &KernelParams<F, N>, ) -> Kernel<F>

pub fn view(&'a self) -> KernelView<'a, F>

Gives a KernelView which has a view on the original kernel’s inner matrix

impl<F: Float> KernelBase<ArrayBase<ViewRepr<&F>, Dim<[usize; 2]>>, CsMatBase<F, usize, &[usize], &[usize], &[F]>>

pub fn to_owned(&self) -> Kernel<F>