Crate linfa_kernel

Kernel methods

Kernel methods are a class of algorithms for pattern analysis, whose best known member is the support vector machine. They owe their name to the kernel functions, which maps the features to some higher-dimensional target space. Common examples for kernel functions are the radial basis function (euclidean distance) or polynomial kernels.

Current State

linfa-kernel currently provides an implementation of kernel methods for RBF and polynomial kernels, with sparse or dense representation. Further a k-neighbour approximation allows to reduce the kernel matrix size.

Low-rank kernel approximation are currently missing, but are on the roadmap. Examples for these are the Nyström approximation or Quasi Random Fourier Features.

Re-exports

pub use inner::Inner;

pub use inner::KernelInner;

Modules

inner

Structs

KernelBase

A generic kernel

KernelParams

Defines the set of parameters needed to build a kernel

Enums

KernelMethod

The inner product definition used by a kernel.

KernelType

Kernel representation, can be either dense or sparse

Type Aliases

Kernel

Type definition of Kernel that owns its inner matrix

KernelView

Type definition of Kernel that borrows its inner matrix