Type Definition linfa_kernel::Kernel

pub type Kernel<F> = KernelBase<Array2<F>, CsMat<F>>;

Type definition of Kernel that owns its inner matrix

Implementations

impl<'a, F: Float> Kernel<F>

pub fn new<N: NearestNeighbour>( dataset: ArrayView2<'a, F>, params: &KernelParams<F, N> ) -> Kernel<F>

pub fn view(&'a self) -> KernelView<'a, F>

Gives a KernelView which has a view on the original kernel’s inner matrix