linfa_kernel

Type Alias KernelView

pub type KernelView<'a, F> = KernelBase<ArrayView2<'a, F>, CsMatView<'a, F>>;

Type definition of Kernel that borrows its inner matrix

Aliased Type

pub struct KernelView<'a, F> {
    pub inner: KernelInner<ArrayBase<ViewRepr<&'a F>, Dim<[usize; 2]>>, CsMatBase<F, usize, &'a [usize], &'a [usize], &'a [F]>>,
    pub method: KernelMethod<<ArrayBase<ViewRepr<&'a F>, Dim<[usize; 2]>> as Inner>::Elem>,
}

Fields
inner: KernelInner<ArrayBase<ViewRepr<&'a F>, Dim<[usize; 2]>>, CsMatBase<F, usize, &'a [usize], &'a [usize], &'a [F]>>method: KernelMethod<<ArrayBase<ViewRepr<&'a F>, Dim<[usize; 2]>> as Inner>::Elem>

The inner product that will be used by the kernel

Implementations

impl<F: Float> KernelView<'_, F>

pub fn to_owned(&self) -> Kernel<F>